requirements.txt
dragonfly2
wxpython
pillow
toml
future
mock
setuptools
win10toast
beautifulsoup4
six
pywin32